Fixing Data Type

+++ McDermott/McDermott.Application/Dtos/Config/CityDto.cs
@@ -2,9 +2,9 @@
 {
     public class CityDto : IMapFrom<City>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         [Required]
-        public int? ProvinceId { get; set; }
+        public long? ProvinceId { get; set; }
 
         [Required]
         [StringLength(200)]
