+++ McDermott/McDermott.Application/Dtos/Employee/DepartmentDto.cs
@@ -2,10 +2,10 @@
 {
     public class DepartmentDto : IMapFrom<Department>
     {
-        public int Id { get; set; }
-        public int? CompanyId { get; set; }
-        public int? ParentDepartmentId { get; set; }
-        public int? ManagerId { get; set; }
+         public long Id { get; set; }
+        public long? CompanyId { get; set; }
+        public long? ParentDepartmentId { get; set; }
+        public long? ManagerId { get; set; }
 
         [StringLength(200)]
         [Required]
