+++ McDermott/McDermott.Application/Dtos/Patient/PatientAllergyDto.cs
@@ -1,9 +1,11 @@
-﻿namespace McDermott.Application.Dtos.Patient
+﻿
+
+namespace McDermott.Application.Dtos.Patient
 {
     public class PatientAllergyDto : IMapFrom<PatientAllergy>
     {
-        public int Id { get; set; }
-        public int UserId { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; }
         public string? Farmacology { get; set; }
         public string? Weather { get; set; }
         public string? Food { get; set; }
