+++ McDermott/McDermott.Application/Dtos/Patient/PatientFamilyRelationDto.cs
@@ -2,10 +2,10 @@
 {
     public class PatientFamilyRelationDto : IMapFrom<PatientFamilyRelation>
     {
-        public int Id { get; set; }
-        public int PatientId { get; set; }
-        public int? FamilyMemberId { get; set; }
-        public int? FamilyId { get; set; }
+         public long Id { get; set; }
+        public long PatientId { get; set; }
+        public long? FamilyMemberId { get; set; }
+        public long? FamilyId { get; set; }
         public string? Relation { get; set; }
 
         public virtual FamilyDto? Family { get; set; }
