+++ McDermott/McDermott.Application/Dtos/Queue/QueueDisplayDto.cs
@@ -8,10 +8,10 @@ namespace McDermott.Application.Dtos.Queue
 {
     public class QueueDisplayDto:IMapFrom<QueueDisplay>
     {
-        public int Id { get; set; }
+         public long Id { get; set; }
         public string? Name { get; set; } = string.Empty;
         public string? NameCounter { get; set; }=string.Empty;
-        public List<int>? CounterId { get; set; } = [];
+        public List<long>? CounterId { get; set; } = [];
         public virtual List<CounterDto>? Counter {  get; set; }
     }
 }
